[DI] Add AutowireRequiredMethodsPass to fix bindings for `@required` methods
