+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/AutowirePassTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\DependencyInjection\Tests\Compiler;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\DependencyInjection\Compiler\AutowireRequiredMethodsPass;
 use Symfony\Component\DependencyInjection\Compiler\AutowirePass;
 use Symfony\Component\DependencyInjection\Compiler\ResolveClassPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -19,6 +20,8 @@ use Symfony\Component\DependencyInjection\Exception\RuntimeException;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\includes\FooVariadic;
 use Symfony\Component\DependencyInjection\TypedReference;
+
+require_once __DIR__.'/../Fixtures/includes/autowiring_classes.php';
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>
@@ -174,7 +177,7 @@ class AutowirePassTest extends TestCase
 
     /**
      * @expectedException \Symfony\Component\DependencyInjection\Exception\AutowiringFailedException
-     * @expectedExceptionMessage Unable to resolve service "private_service": constructor of class "Symfony\Component\DependencyInjection\Tests\Compiler\PrivateConstructor" must be public.
+     * @expectedExceptionMessage Invalid service "private_service": constructor of class "Symfony\Component\DependencyInjection\Tests\Compiler\PrivateConstructor" must be public.
      */
     public function testPrivateConstructorThrowsAutowireException()
     {
@@ -554,6 +557,7 @@ class AutowirePassTest extends TestCase
         ;
 
         (new ResolveClassPass())->process($container);
+        (new AutowireRequiredMethodsPass())->process($container);
         (new AutowirePass())->process($container);
 
         $methodCalls = $container->getDefinition('setter_injection')->getMethodCalls();
@@ -590,6 +594,7 @@ class AutowirePassTest extends TestCase
         ;
 
         (new ResolveClassPass())->process($container);
+        (new AutowireRequiredMethodsPass())->process($container);
         (new AutowirePass())->process($container);
 
         $methodCalls = $container->getDefinition('setter_injection')->getMethodCalls();
@@ -687,6 +692,8 @@ class AutowirePassTest extends TestCase
         $aDefinition = $container->register('setter_injection_collision', SetterInjectionCollision::class);
         $aDefinition->setAutowired(true);
 
+        (new AutowireRequiredMethodsPass())->process($container);
+
         $pass = new AutowirePass();
         $pass->process($container);
     }
@@ -770,6 +777,7 @@ class AutowirePassTest extends TestCase
         }
 
         (new ResolveClassPass())->process($container);
+        (new AutowireRequiredMethodsPass())->process($container);
         (new AutowirePass())->process($container);
     }
 
@@ -778,7 +786,7 @@ class AutowirePassTest extends TestCase
         return array(
             array('setNotAutowireable', 'Cannot autowire service "foo": argument "$n" of method "Symfony\Component\DependencyInjection\Tests\Compiler\NotWireable::setNotAutowireable()" has type "Symfony\Component\DependencyInjection\Tests\Compiler\NotARealClass" but this class cannot be loaded.'),
             array('setDifferentNamespace', 'Cannot autowire service "foo": argument "$n" of method "Symfony\Component\DependencyInjection\Tests\Compiler\NotWireable::setDifferentNamespace()" references class "stdClass" but no such service exists. It cannot be auto-registered because it is from a different root namespace.'),
-            array(null, 'Cannot autowire service "foo": method "Symfony\Component\DependencyInjection\Tests\Compiler\NotWireable::setProtectedMethod()" must be public.'),
+            array(null, 'Invalid service "foo": method "Symfony\Component\DependencyInjection\Tests\Compiler\NotWireable::setProtectedMethod()" must be public.'),
         );
     }
 
@@ -857,335 +865,5 @@ class AutowirePassTest extends TestCase
 
         $pass = new AutowirePass();
         $pass->process($container);
-    }
-}
-
-class Foo
-{
-}
-
-class Bar
-{
-    public function __construct(Foo $foo)
-    {
-    }
-}
-
-interface AInterface
-{
-}
-
-class A implements AInterface
-{
-    public static function create(Foo $foo)
-    {
-    }
-}
-
-class B extends A
-{
-}
-
-class C
-{
-    public function __construct(A $a)
-    {
-    }
-}
-
-interface DInterface
-{
-}
-
-interface EInterface extends DInterface
-{
-}
-
-interface IInterface
-{
-}
-
-class I implements IInterface
-{
-}
-
-class F extends I implements EInterface
-{
-}
-
-class G
-{
-    public function __construct(DInterface $d, EInterface $e, IInterface $i)
-    {
-    }
-}
-
-class H
-{
-    public function __construct(B $b, DInterface $d)
-    {
-    }
-}
-
-class D
-{
-    public function __construct(A $a, DInterface $d)
-    {
-    }
-}
-
-class E
-{
-    public function __construct(D $d = null)
-    {
-    }
-}
-
-class J
-{
-    public function __construct(I $i)
-    {
-    }
-}
-
-interface CollisionInterface
-{
-}
-
-class CollisionA implements CollisionInterface
-{
-}
-
-class CollisionB implements CollisionInterface
-{
-}
-
-class CannotBeAutowired
-{
-    public function __construct(CollisionInterface $collision)
-    {
-    }
-}
-
-class CannotBeAutowiredForwardOrder
-{
-    public function __construct(CollisionA $a, CollisionInterface $b, CollisionB $c)
-    {
-    }
-}
-
-class CannotBeAutowiredReverseOrder
-{
-    public function __construct(CollisionA $a, CollisionB $c, CollisionInterface $b)
-    {
-    }
-}
-
-class Lille
-{
-}
-
-class Dunglas
-{
-    public function __construct(Lille $l)
-    {
-    }
-}
-
-class LesTilleuls
-{
-    public function __construct(Dunglas $j, Dunglas $k)
-    {
-    }
-}
-
-class OptionalParameter
-{
-    public function __construct(CollisionInterface $c = null, A $a, Foo $f = null)
-    {
-    }
-}
-
-class BadTypeHintedArgument
-{
-    public function __construct(Dunglas $k, NotARealClass $r)
-    {
-    }
-}
-class BadParentTypeHintedArgument
-{
-    public function __construct(Dunglas $k, OptionalServiceClass $r)
-    {
-    }
-}
-class NotGuessableArgument
-{
-    public function __construct(Foo $k)
-    {
-    }
-}
-class NotGuessableArgumentForSubclass
-{
-    public function __construct(A $k)
-    {
-    }
-}
-class MultipleArguments
-{
-    public function __construct(A $k, $foo, Dunglas $dunglas)
-    {
-    }
-}
-
-class MultipleArgumentsOptionalScalar
-{
-    public function __construct(A $a, $foo = 'default_val', Lille $lille = null)
-    {
-    }
-}
-class MultipleArgumentsOptionalScalarLast
-{
-    public function __construct(A $a, Lille $lille, $foo = 'some_val')
-    {
-    }
-}
-class MultipleArgumentsOptionalScalarNotReallyOptional
-{
-    public function __construct(A $a, $foo = 'default_val', Lille $lille)
-    {
-    }
-}
-
-/*
- * Classes used for testing createResourceForClass
- */
-class ClassForResource
-{
-    public function __construct($foo, Bar $bar = null)
-    {
-    }
-
-    public function setBar(Bar $bar)
-    {
-    }
-}
-class IdenticalClassResource extends ClassForResource
-{
-}
-
-class ClassChangedConstructorArgs extends ClassForResource
-{
-    public function __construct($foo, Bar $bar, $baz)
-    {
-    }
-}
-
-class SetterInjection extends SetterInjectionParent
-{
-    /**
-     * @required
-     */
-    public function setFoo(Foo $foo)
-    {
-        // should be called
-    }
-
-    /** @inheritdoc*/
-    public function setDependencies(Foo $foo, A $a)
-    {
-        // should be called
-    }
-
-    /** {@inheritdoc} */
-    public function setWithCallsConfigured(A $a)
-    {
-        // this method has a calls configured on it
-    }
-
-    public function notASetter(A $a)
-    {
-        // should be called only when explicitly specified
-    }
-
-    /**
-     * @required*/
-    public function setChildMethodWithoutDocBlock(A $a)
-    {
-    }
-}
-
-class SetterInjectionParent
-{
-    /** @required*/
-    public function setDependencies(Foo $foo, A $a)
-    {
-        // should be called
-    }
-
-    public function notASetter(A $a)
-    {
-        // @required should be ignored when the child does not add @inheritdoc
-    }
-
-    /**	@required <tab> prefix is on purpose */
-    public function setWithCallsConfigured(A $a)
-    {
-    }
-
-    /** @required */
-    public function setChildMethodWithoutDocBlock(A $a)
-    {
-    }
-}
-
-class SetterInjectionCollision
-{
-    /**
-     * @required
-     */
-    public function setMultipleInstancesForOneArg(CollisionInterface $collision)
-    {
-        // The CollisionInterface cannot be autowired - there are multiple
-
-        // should throw an exception
-    }
-}
-
-class NotWireable
-{
-    public function setNotAutowireable(NotARealClass $n)
-    {
-    }
-
-    public function setBar()
-    {
-    }
-
-    public function setOptionalNotAutowireable(NotARealClass $n = null)
-    {
-    }
-
-    public function setDifferentNamespace(\stdClass $n)
-    {
-    }
-
-    public function setOptionalNoTypeHint($foo = null)
-    {
-    }
-
-    public function setOptionalArgNoAutowireable($other = 'default_val')
-    {
-    }
-
-    /** @required */
-    protected function setProtectedMethod(A $a)
-    {
-    }
-}
-
-class PrivateConstructor
-{
-    private function __construct()
-    {
     }
 }
